# [ENH] Faster histogram