--- a/mdtraj/geometry/include/numericutils.h
+++ b/mdtraj/geometry/include/numericutils.h
@@ -1,0 +1,12 @@
+#ifndef _GEOMETRY_H_
+#define _GEOMETRY_H_
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int histogram(const float* a, const float* bin_edges, float* out);
+
+#ifdef __cplusplus
+}
+#endif
+#endif