--- a/mdtraj/geometry/include/numericutils.h
+++ b/mdtraj/geometry/include/numericutils.h
@@ -4,7 +4,9 @@
 extern "C" {
 #endif
 
-int histogram(const float* a, const float* bin_edges, float* out);
+int histogram(const float* data, const int n_data, const float* bins,
+              const int n_bins, const float min_bin, const float max_bin,
+              int* out);
 
 #ifdef __cplusplus
 }