--- a/mdtraj/geometry/rdf.py
+++ b/mdtraj/geometry/rdf.py
@@ -29,6 +29,7 @@ import numpy as np
 
 from mdtraj.utils import ensure_type
 from mdtraj.geometry.distance import compute_distances
+from mdtraj.geometry import _numericutils
 
 
 __all__ = ['compute_rdf']
@@ -90,7 +91,10 @@ def compute_rdf(traj, pair_names=None, r_range=None, bin_width=0.005,
     bins = np.arange(r_range[0], r_range[1], bin_width)
 
     distances = compute_distances(traj, pairs, periodic=periodic, opt=opt)
-    g_r, edges = np.histogram(distances, bins=bins)
+    if opt:
+        g_r, edges = _numericutils._histogram(distances, bins=bins)
+    else:
+        g_r, edges = np.histogram(distances, bins=bins)
     r = 0.5 * (edges[1:] + edges[:-1])
 
     # Normalize.