The current RDF code spends ~80% of its time in the histogram on my laptop so I figured I'd try my hand at writing a faster one (is this surprising/suspicious or is your distance code just that awesome?).

From some quick timings, this histogram is a ~10x faster for the RDF use case:
```python
In [8]: a = np.random.rand(10000000)

In [9]:  %timeit np.histogram(a, bins=100)
1 loops, best of 3: 928 ms per loop

In [10]: %timeit _numericutils._histogram(a, bins=100)
10 loops, best of 3: 83.6 ms per loop

In [11]: a = np.random.rand(100000000)

In [12]:  %timeit np.histogram(a, bins=100)
1 loops, best of 3: 8.91 s per loop

In [13]: %timeit _numericutils._histogram(a, bins=100)
1 loops, best of 3: 975 ms per loop
```

That being said, this is my first time dabbling with cython and I don't write a lot of C/C++ so I've got a few questions:
* First and foremost, do you want something like this in your code base?
* If yes, do my file structure and function names make sense to you? Any conventions that I'm doing a terrible injustice to?
* I've tried to reproduce the behavior of the numpy histogram with the exception of renaming the `range` arg to `bin_range`. The PR also includes all of the unit tests that numpy uses but modified to also check that the results are equal to those produced by numpy. Any objections to this approach and/or that particular change?

Other notes/TODO's:
* In `test_f32_rounding()` There is a slight difference (out to ~4-5 decimals) in the returned edges as compared to numpy. Still need to identify where this is coming from and fix it. The actual values in the histogram appear to be consistent, however, for the tests I've run so far.
* I have yet to implement the numpy functionality provided by the args `normed`, `weights` and `density` because there wasn't an immediate need.
* The current code could easily just be regular cython without further wrapping but I figured it may be worthwhile to keep the option open for some kind of SIMD acceleration (which is a bit [tricky for histograms](http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.80.9041&rep=rep1&type=pdf)). Might also make it easier to add some OpenMP threading to accumulate sub-histograms but that can be done in cython too (some potential [code](https://github.com/kif/pyFAI/blob/master/src/histogram_omp.pyx) to borrow from).