@kyleabeauchamp 2015-02-23
I guess my first question is whether this has already been implemented by, say, scipy, sklearn, or pandas, which might be considered for inclusion as an optional import.