@ctk3b 2015-02-23
[Scipy uses numpy internally](https://github.com/scipy/scipy/blob/v0.14.0/scipy/stats/stats.py#L1666)
[Pandas has some cython](https://github.com/pydata/pandas/blob/9f439f0e53409fd48e5486a3acf84b844f200270/pandas/hashtable.pyx#L983)

I can't immediately find anything in sklearn. I'll benchmark the pandas one.