@kyleabeauchamp 2015-02-23
Yeah, I thought the pandas folks have made some performance improvements over numpy / scipy for several "mission critical" tasks.  

Also, we already have some optional pandas dependencies scattered around the repo, so it wouldn't be a big burden to use that if it's fast.   

Finally, if your improved histogramming is a major performance improvement, I wonder if you're better off filing a PR directly to pandas.  Certainly, that would impact more users.