@rmcgibbo 2015-02-23
> The current RDF code spends ~80% of its time in the histogram on my laptop 

How does this compare to the amount of time it takes, for example, to simply load the trajectory files from disk? Is this a big overall contributor to the time it takes to accomplish some task, or are we talking about accelerating 80% of something that already only takes milliseconds?