@rmcgibbo 2015-02-23
The pandas histogram uses a hash table I think. Different implementation from numpy. At one point, I believe there was a discussion on numpy-dev or scipy-dev about this.