@ctk3b 2015-02-23
@kyleabeauchamp  Well I haven't really done anything earth shattering here - it's a pretty bare bones implementation of only the part I wanted for the RDF. The pandas implementation doesn't seem to be any faster than numpy so maybe it'd be worth looking into improving their cython code.

@rmcgibbo for the tip3p trajectory, I get ~14s total, 11 spent in np.histogram, 2.5 in distances. With the PR histogram it's 4.1 total, 2.7 in distances and 1.0 in the histogram (note these are based on just a single run but I think the loading isn't limiting in either case).