@ctk3b 2015-02-23
Using: 
```python
import cProfile, pstats, StringIO
import mdtraj as md
from mdtraj.testing import get_fn
pr = cProfile.Profile()
pr.enable()

traj = md.load(get_fn('tip3p_300K_1ATM.xtc'), top=get_fn('tip3p_300K_1ATM.pdb'))
r, g_r = md.compute_rdf(traj)

pr.disable()
s = StringIO.StringIO()
sortby = 'cumulative'
ps = pstats.Stats(pr, stream=s).sort_stats(sortby)
ps.print_stats()
print s.getvalue()
```