@ctk3b 2015-02-23
Semi-related, how do you guys test across python versions? I'm using conda environments but from my understanding those don't play nicely with tox.