@kyleabeauchamp 2015-02-23
Personally, I let the travis grid handle python versions for the most part.  When I need to test things locally, I end up using either conda environments or a separate (clean) miniconda install.