@ctk3b 2015-02-24
Also if you guys aren't comfortable adding a re-invented wheel to your repo, that's more than understandable - it really does seem like it would belong in a package like numpy or pandas. This mostly started as a learning experience to fully understand how all the cython extensions work in mdtraj. 

I'll take a closer look at how accelerating the pandas code would work.