@kyleabeauchamp 2015-02-24
I would look up the discussion on the listserv first, in case there are
some hard problems that we're not anticipating.
On Feb 23, 2015 7:11 PM, "Christoph Klein" <notifications@github.com> wrote:

> Also if you guys aren't comfortable adding a re-invented wheel to your
> repo, that's more than understandable - it really does seem like it would
> belong in a package like numpy or pandas. This mostly started as a learning
> experience to fully understand how all the cython extensions work in
> mdtraj.
>
> I'll take a closer look at how accelerating the pandas code would work.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/mdtraj/mdtraj/pull/734#issuecomment-75669092>.
>