@rmcgibbo 2015-02-24
Overall, I'm a little -1 for the following reasons:

1. In general, I think we should avoid having code in mdtraj that isn't "core" to our mission and competencies. We want to lean on numpy and friends as much as possible. They spend a lot of effort, which we don't have the manpower (or will) to reproduce, dealing with cross-platform compatibility, documentation, etc.

2. I think we should have a particularly high bar for including new C or C++ code in mdtraj. Two reasons:
  - A smaller percentage of the mdtraj developers or people who might in the future become mdtraj developers count hacking in C among their skills. The pool of people who can identify and fix bugs, do maintenance, and extend the code are simply smaller. This is worrisome if we want the code to continue to be maintained through a cycle of core-developer-graduates-and-a-new-team-steps-up.  Honestly, this is something I'm worried about already.
 - Cross platform support in C and C++ is just harder. Especially with C and Windows. There are pretty substantial differences between what's supported in C with MSVC and with gcc/clang. And the pool of developers who can work with windows and C/C++ is pretty small for us.

Obviously these considerations have to be balanced against performance. There is a balancing act. I'm not sure in general how to find the right balance, and we don't really have a project protocol laid out of this kind of thing.