@kyleabeauchamp 2015-02-24
I agree that we need a high bar for additional C code.  

My vote is to first see whether this belongs in Numpy / pandas--in those places, it's easier to "hand off" code to the community.  If we merge it here, it means that we personally are going to be maintaining it.