@ctk3b 2015-02-24
That all sounds very reasonable. I'll do some closer reading of the numpy and pandas lists to see what exactly people are working on. Specifically for the RDF, we don't really need a feature rich implementation but as mentioned, maintaining it across platforms will be the tricky part.

I was really surprised though that something like this doesn't exist already. It sounded so simple and wasn't hard to write but I am only looking at a pretty specific use case. From what I've been reading on the numpy PR's that worked on histogram performance, the main concern seems to be generality for all inputs sizes.