@swails 2015-02-24
I'll interject here with a slightly different perspective, particularly with regards to:

> Obviously these considerations have to be balanced against performance. There is a balancing act. I'm not sure in general how to find the right balance, and we don't really have a project protocol laid out of this kind of thing.

I think a good philosophy is to have a software project like this be, first and foremost, useful to the people that develop it.  With that in mind, @ctk3b, how critical to your work is the RDF speedup?  In particular:

1. Does it make the difference between using MDTraj vs. a different package for your application?
2. Does your new code eliminate performance-based limits on the neat stuff you're trying to do with it, allowing you to do more, neater stuff? (related to 1, really)
3. Does this need to be part of MDTraj proper for it to be realistically useful to you? (i.e., can you use it in your own work as a kind of 'plug-in' type of thing?)

The answers to these questions may help define where the importance of this feature lies in the balancing act @rmcgibbo was referring to.

Relating to 3 -- perhaps another route you can go is to provide an option to register a histogramming function to use with the RDF, defaulting to numpy's implementation.  That way, you can play around with which one is best, and plug in your own if none of the available ones suffice, without having to dump anything like this into MDTraj proper.  Just a thought.