@ctk3b 2015-07-21
Huzzah! The full RDF looks to be about ~3-4 times faster for this particular system (water box) which is consistent with @gabrielelanaro's solution in chemlab which also uses bincount.

However, it requires a small tweak to either the RDF code or alternatively the PR which may or may not be acceptable:

This [line](https://github.com/numpy/numpy/pull/6100/files#diff-9d44603693b9544b900e135a203b3ae2R188) in the PR checks to see if "the bins are equal width" by checking if `bins` is not iterable. Right now I set up the RDF code to just always provide an array of equal width bins to the histogram function. They're still equal width but the variable is an array and hence iterable. The easiest tweak is probably just to change how the RDF code works, but on the other hand, the numpy code is not technically checking for a case where the bins are equal width.

Timeit results:
```python
# 1.10.0.dev0+daf0497
In [2]: %timeit run() 
1 loops, best of 3: 10.1 s per loop

# #6100
In [3]: %timeit run()
1 loops, best of 3: 2.71 s per loop
```

And the performance breakdown with relevant parts shown:
New and improved:

```
85663 function calls (83563 primitive calls) in 2.769 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.046    0.046    2.769    2.769 test_rdf.py:4(run)
        1    0.000    0.000    2.596    2.596 //anaconda/envs/hist2/lib/python3.4/site-packages/mdtraj/geometry/rdf.py:33(compute_rdf)
        1    0.001    0.001    1.631    1.631 //anaconda/envs/hist2/lib/python3.4/site-packages/mdtraj/geometry/distance.py:42(compute_distances)
        1    1.628    1.628    1.628    1.628 {_geometry._dist_mic}
        1    0.731    0.731    0.960    0.960 //anaconda/envs/hist2/lib/python3.4/site-packages/numpy/lib/function_base.py:78(histogram)
     1831    0.097    0.000    0.097    0.000 {built-in method bincount}
        1    0.000    0.000    0.071    0.071 //anaconda/envs/hist2/lib/python3.4/site-packages/mdtraj/core/trajectory.py:294(load)
```

Old and totes unimproved:

```
130412 function calls (128312 primitive calls) in 10.081 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.048    0.048   10.081   10.081 test_rdf.py:4(run)
        1    0.000    0.000    9.907    9.907 //anaconda/envs/hist/lib/python3.4/site-packages/mdtraj/geometry/rdf.py:33(compute_rdf)
        1    0.056    0.056    8.263    8.263 //anaconda/envs/hist/lib/python3.4/site-packages/numpy/lib/function_base.py:78(histogram)
     1831    0.008    0.000    7.897    0.004 //anaconda/envs/hist/lib/python3.4/site-packages/numpy/core/fromnumeric.py:712(sort)
     1833    7.790    0.004    7.790    0.004 {method 'sort' of 'numpy.ndarray' objects}
        1    0.002    0.002    1.637    1.637 //anaconda/envs/hist/lib/python3.4/site-packages/mdtraj/geometry/distance.py:42(compute_distances)
        1    1.633    1.633    1.633    1.633 {_geometry._dist_mic}
     3662    0.207    0.000    0.207    0.000 {method 'searchsorted' of 'numpy.ndarray' objects}
```




Testing script:
```python
import mdtraj as md
from mdtraj.testing import get_fn

def run():
    traj = md.load(get_fn('tip3p_300K_1ATM.xtc'), top=get_fn('tip3p_300K_1ATM.pdb'))
    pairs = traj.top.select_pairs('all', 'all')
    r, g_r = md.compute_rdf(traj, pairs)

if __name__ == '__main__':
    import cProfile, pstats, io
    pr = cProfile.Profile()
    pr.enable()

    run()

    pr.disable()
    s = io.StringIO()
    sortby = 'cumulative'
    ps = pstats.Stats(pr, stream=s).sort_stats(sortby)
    ps.print_stats()
    print(s.getvalue())
```